+++ dune/localfunctions/hierarchical.hh
@@ -1,0 +1,9 @@
+// -*- tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 2 -*-
+// vi: set et ts=4 sw=2 sts=2:
+/** \file
+    \brief Convenience header that includes all available hierarchical LocalFiniteElements
+ */
+
+#include <dune/localfunctions/hierarchical/hierarchicalp2.hh>
+#include <dune/localfunctions/hierarchical/hierarchicalp2withelementbubble.hh>
+#include <dune/localfunctions/hierarchical/hierarchicalprismp2.hh>